requirements.txt
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
boto3==1.9.167
botocore==1.12.167
certifi==2019.3.9
chardet==3.0.4
chardet2==2.0.3
Click==7.0
construct==2.5.3
coverage==4.5.3
cssselect==1.0.3
cycler==0.10.0
ddt==1.2.1
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-ckeditor==5.7.1
django-cleanup==3.2.0
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-embed-video==1.2.0
django-filters==0.2.1
django-honeypot==0.7.0
django-imagekit==4.0.2
django-jalali-date==0.2.9
django-js-asset==1.2.2
django-markupfield==1.5.1
django-memcache-status==2.1
django-mongodb-engine==0.6.0
django-pipeline==1.6.14
django-rest-framework==0.1.0
django-sitetree==1.13.0
django-storages==1.7.1
django-waffle==0.16.0
djangorestframework==3.9.4
docutils==0.14
elasticsearch==7.0.2
entrypoints==0.3
factory-boy==2.12.0
Faker==1.0.7
feedparser==5.2.1
Flask==1.0.3
future==0.17.1
gunicorn==19.9.0
haystack==0.42
httplib2==0.13.0
icalendar==4.0.3
idna==2.8
imagesize==1.1.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jdatetime==3.2.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kavenegar==1.1.2
kiwisolver==1.1.0
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
model-mommy==1.6.0
mysqlclient==1.4.2.post1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
oauthlib==3.0.1
packaging==19.0
pandocfilters==1.4.2
parso==0.4.0
pefile==2019.4.18
pexpect==4.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==6.0.0
powerline-status==2.7
prometheus-client==0.7.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pymongo==3.8.0
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
python-ptrace==0.9.3
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.1
raven==6.10.0
requests==2.22.0
requests-oauthlib==1.2.0
response==0.3.0
s3transfer==0.2.1
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9.1
Sphinx==2.1.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.4
sqlparse==0.3.0
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
testypie==1.0.0
text-unidecode==1.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whitenoise==4.1.2
widgetsnbextension==3.4.2